rename registry's 'api' package to 'server'
