+++ controller/client/client.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/Sirupsen/logrus"
 	"github.com/amalgam8/amalgam8/controller/rules"
-	"github.com/amalgam8/amalgam8/registry/api/env"
+	"github.com/amalgam8/amalgam8/registry/server/env"
 )
 
 const defaultTimeout = 30 * time.Second
