+++ registry/server/middleware/auth_test.go
