+++ registry/server/middleware/context.go
@@ -19,7 +19,7 @@ import (
 
 	"github.com/ant0ine/go-json-rest/rest"
 
-	"github.com/amalgam8/amalgam8/registry/api/env"
+	"github.com/amalgam8/amalgam8/registry/server/env"
 )
 
 // ContextMiddleware initializes a context and puts it in the request's env map
