+++ registry/server/middleware/secure.go
