+++ registry/server/middleware/secure_test.go
