+++ registry/server/middleware/trace.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/ant0ine/go-json-rest/rest"
 
-	"github.com/amalgam8/amalgam8/registry/api/env"
+	"github.com/amalgam8/amalgam8/registry/server/env"
 )
 
 var globalTraceNameGenerator = NewNameGenerator(defaultDictionary)
