+++ registry/server/protocol/amalgam8/endpoints.go
