+++ registry/server/protocol/amalgam8/instance.go
