+++ registry/server/protocol/amalgam8/instance_address.go
