+++ registry/server/protocol/amalgam8/instance_handler.go
@@ -25,8 +25,8 @@ import (
 	log "github.com/Sirupsen/logrus"
 	"github.com/ant0ine/go-json-rest/rest"
 
-	"github.com/amalgam8/amalgam8/registry/api/env"
-	"github.com/amalgam8/amalgam8/registry/api/middleware"
+	"github.com/amalgam8/amalgam8/registry/server/env"
+	"github.com/amalgam8/amalgam8/registry/server/middleware"
 	"github.com/amalgam8/amalgam8/registry/store"
 	"github.com/amalgam8/amalgam8/registry/utils/i18n"
 	"github.com/amalgam8/amalgam8/registry/utils/reflection"
