+++ registry/server/protocol/amalgam8/instance_links.go
