+++ registry/server/protocol/amalgam8/instances_list.go
