+++ registry/server/protocol/amalgam8/service_instance_list.go
