+++ registry/server/protocol/amalgam8/services_list.go
