+++ registry/server/protocol/eureka/amazon_dcinfo.go
