+++ registry/server/protocol/eureka/applications.go
