+++ registry/server/protocol/eureka/applications_list.go
