+++ registry/server/uptime/endpoints.go
