+++ registry/server/uptime/routes.go
